requirements.txt
affine==2.3.0
attrs==21.2.0
backcall==0.2.0
Cartopy==0.20.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
contextily==1.2.0
cycler==0.11.0
decorator==5.1.0
Fiona==1.8.20
geographiclib==1.52
geopandas==0.10.2
geopy==2.2.0
idna==3.3
ipython==7.30.1
jedi==0.18.1
jellyfish==0.6.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
matplotlib-inline==0.1.3
mercantile==1.2.1
munch==2.5.0
numpy==1.21.4
oauthlib==3.1.1
pandas==1.3.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.10.0
pyparsing==3.0.4
pyproj==3.2.1
pyshp==2.1.3
python-dateutil==2.8.2
pytz==2021.3
rasterio==1.2.10
requests==2.26.0
requests-oauthlib==1.3.0
Shapely==1.8.0
six==1.16.0
snuggs==1.4.7
traitlets==5.1.1
tweepy==4.3.0
urllib3==1.26.7
us==2.0.2
wcwidth==0.2.5
xyzservices==2021.11.0
